move data karyawan from components to containers folder
