+++ src/layouts/AdminLayout/AdminLayout.tsx
@@ -2,7 +2,6 @@ import React, {
   PropsWithChildren, useCallback, useEffect, useState,
 } from 'react'
 import { useResizeDetector } from 'react-resize-detector'
-import Head from 'next/head'
 import Sidebar, { SidebarOverlay } from '@layouts/AdminLayout/Sidebar/Sidebar'
 import Header from '@layouts/AdminLayout/Header/Header'
 import Footer from '@layouts/AdminLayout/Footer/Footer'
@@ -44,27 +43,21 @@ export default function AdminLayout({ children }: PropsWithChildren) {
   }, [setIsShowSidebarMd])
 
   return (
-      <>
-        <Head>
-          <title>Mar Fashion Web</title>
-          <meta name="description" content="Generated by create next app" />
-          <link rel="icon" href="/favicon.ico" />
-        </Head>
-
-        <div ref={ref} className="position-absolute w-100" />
-          <Sidebar isShow={isShowSidebar} isShowMd={isShowSidebarMd} />
-        
-        <div className="wrapper d-flex flex-column min-vh-100 bg-light">
-          <Header toggleSidebar={toggleIsShowSidebar} toggleSidebarMd={toggleIsShowSidebarMd} />
-          <div className="body flex-grow-1 px-3">
-            <Container fluid="lg">
-              {children}
-            </Container>
-          </div>
-          <Footer />
+    <>
+      <div ref={ref} className="position-absolute w-100" />
+        <Sidebar isShow={isShowSidebar} isShowMd={isShowSidebarMd} />
+      
+      <div className="wrapper d-flex flex-column min-vh-100 bg-light">
+        <Header toggleSidebar={toggleIsShowSidebar} toggleSidebarMd={toggleIsShowSidebarMd} />
+        <div className="body flex-grow-1 px-3">
+          <Container fluid="lg">
+            {children}
+          </Container>
         </div>
+        <Footer />
+      </div>
 
-        <SidebarOverlay isShowSidebar={isShowSidebar} toggleSidebar={toggleIsShowSidebar} />
-      </>
+      <SidebarOverlay isShowSidebar={isShowSidebar} toggleSidebar={toggleIsShowSidebar} />
+    </>
   )
 }
